@@ -29,15 +29,15 @@
     "antd": "^5.3.1",
     "array-move": "^4.0.0",
     "openai": "^3.2.1",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "tailwindcss": "^3.2.7"
   },
   "devDependencies": {
     "@types/chrome": "0.0.158",
     "@types/jest": "^27.0.2",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
+    "@types/react": "^18.2.0",
+    "@types/react-dom": "^18.2.0",
     "copy-webpack-plugin": "^9.0.1",
     "css-loader": "^6.7.3",
     "glob": "^7.1.6",
